+++ src/realm/sync/noinst/client_impl_base.cpp
@@ -1595,9 +1595,6 @@ void Session::send_message()
     if (!m_bind_message_sent)
         return send_bind_message(); // Throws
 
-    if (!m_access_token_sent)
-        return send_refresh_message(); // Throws
-
     if (!m_ident_message_sent) {
         if (have_client_file_ident())
             send_ident_message(); // Throws
@@ -1659,7 +1656,6 @@ void Session::send_bind_message()
     m_conn.initiate_write_message(out, this);                           // Throws
 
     m_bind_message_sent = true;
-    m_access_token_sent = true;
 
     // Ready to send the IDENT message if the file identifier pair is already
     // available.
@@ -1891,37 +1887,6 @@ void Session::send_mark_message()
 }
 
 
-void Session::send_refresh_message()
-{
-    REALM_ASSERT(m_state == Active);
-    REALM_ASSERT(m_bind_message_sent);
-    REALM_ASSERT(!m_unbind_message_sent);
-    REALM_ASSERT(!m_access_token_sent);
-
-    // Discard the token since it's ignored by the server.
-    std::string empty_access_token{};
-    std::size_t signed_access_token_size = empty_access_token.size();
-
-    logger.debug("Sending: REFRESH(signed_user_token_size=%1)",
-                 signed_access_token_size); // Throws
-
-    ClientProtocol& protocol = m_conn.get_client_protocol();
-    OutputBuffer& out = m_conn.get_output_buffer();
-    session_ident_type session_ident = get_ident();
-    protocol.make_refresh_message(out, session_ident, empty_access_token); // Throws
-    m_conn.initiate_write_message(out, this);                              // Throws
-
-    m_access_token_sent = true;
-
-    // If the IDENT message has not yet been sent, is now ready to be sent if the
-    // file identifier pair has become available. If IDENT message has been
-    // sent, various other messages may be waiting to be sent, but in that case,
-    // we also have the file identifier pair.
-    if (have_client_file_ident())
-        enlist_to_send(); // Throws
-}
-
-
 void Session::send_unbind_message()
 {
     REALM_ASSERT(m_state == Deactivating || m_error_message_received);
@@ -1969,7 +1934,7 @@ std::error_code Session::receive_ident_message(SaltedFileIdent client_file_ident
     m_client_file_ident = client_file_ident;
 
     if (REALM_UNLIKELY(get_client().is_dry_run())) {
-        // Ready to send the IDENT (or REFRESH) message
+        // Ready to send the IDENT message
         ensure_enlisted_to_send(); // Throws
         return std::error_code{};  // Success
     }
@@ -2036,7 +2001,7 @@ std::error_code Session::receive_ident_message(SaltedFileIdent client_file_ident
         this->m_progress.upload.client_version = 0;
     }
 
-    // Ready to send the IDENT (or REFRESH) message
+    // Ready to send the IDENT message
     ensure_enlisted_to_send(); // Throws
     return std::error_code{};  // Success
 }